add: add new feature

+++ apps/pedagog/signals/signals.py
@@ -13,6 +13,7 @@ from apps.websocket.models.notification import Notification
 from apps.pedagog.models.media import Media
 from apps.pedagog.models.converted_media import ConvertedMedia
 from apps.pedagog.tasks.convert_file import convert_image_create
+from apps.pedagog.tasks.preview import save_preview
 
 from apps.shared.utils.convert_image import convert_pdf_to_images, convert_pptx_to_images, convert_docx_to_images, add_multiple_icons_to_image, convert_office_to_pdf
 
@@ -92,9 +93,13 @@ def file_status_pre_save(sender, instance, **kwargs):
 
 @receiver(post_save, sender=Media)
 def convert_image_on_save_media(sender, instance, created, **kwargs):
-    media_types = ConvertedMedia.objects.filter(media=instance.id)
-    if media_types:
-        for media in media_types:
-            media.delete()
+    if instance.type in ['mp3', 'mp4']:
+        save_preview.delay(instance.id)
+    else:
+        media_types = ConvertedMedia.objects.filter(media=instance.id)
+        if media_types:
+            for media in media_types:
+                media.delete()
+        
+        convert_image_create.delay(instance.id)
     
-    convert_image_create.delay(instance.id)
